Account for future versions of required packages.
We know that django-saltapi works with at least the specified
versions. It is very likely that it will work with previous versions
of Django, but it is untested for now. Salt is more of a moving target
so it is unlikely that anyone has a requirement to run earlier
versions of it. For example Ubuntu 12.10 has Salt version 0.10.1,
which is ancient measured in Salted years.